Bug Fix: Jumping Morecast Map (#1338)
- Pushes VectorLayer child creation down to the Map component so WeatherMap (parent) doesn't re-render it with the BC center point, every time station selection changes
- Pulls out VectorLayer-only logic into FireIndicesVectorLayer to simplify the Map component
